move some more files
