--- app/src/androidMain/kotlin/de/westnordost/streetcomplete/data/DbModule.kt
@@ -1,11 +1,0 @@
-package de.westnordost.streetcomplete.data
-
-import de.westnordost.streetcomplete.ApplicationConstants
-import org.koin.dsl.module
-
-val dbModule = module {
-    single<Database> {
-        val sqLite = StreetCompleteSQLiteOpenHelper(get(), ApplicationConstants.DATABASE_NAME)
-        AndroidDatabase(sqLite.writableDatabase)
-    }
-}
