+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/AllEditTypesModule.kt
