+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/Cleaner.kt
@@ -13,6 +13,7 @@ import de.westnordost.streetcomplete.util.logs.Log
 import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.IO
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.launch
