+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/osm/osmquests/OsmQuestModule.kt
@@ -1,6 +1,5 @@
 package de.westnordost.streetcomplete.data.osm.osmquests
 
-import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
 val osmQuestModule = module {
@@ -8,6 +7,4 @@ val osmQuestModule = module {
     factory { OsmQuestsHiddenDao(get()) }
 
     single<OsmQuestSource> { get<OsmQuestController>() }
-
-    single { OsmQuestController(get(), get(), get(), get(), get(named("CountryBoundariesLazy"))) }
 }
